accountManager/requirements.txt
Flask
flask_cors
requests
python-dotenv